Switching resource variables to int, removed oil processing, needs test

+++ src/com/milesseventh/wargames/ResourceStorage.java
@@ -2,32 +2,32 @@ package com.milesseventh.wargames;
 
 public class ResourceStorage {
 	public String name;
-	private float[] resources = {0, 0, 0, 0, 0};
+	private int[] resources = {0, 0, 0, 0, 0};
 	
 	public ResourceStorage(String nname) {
 		name = nname;
 	}
 	
-	public float sum(){
-		float r = 0;
-		for (float i: resources)
+	public int sum(){
+		int r = 0;
+		for (int i: resources)
 			r += i;
 		return r;
 	}
 
-	public void add(Resource r, float size){
+	public void add(Resource r, int size){
 		resources[r.ordinal()] += size;
 	}
 	
-	public float get(Resource r){
+	public int get(Resource r){
 		return resources[r.ordinal()];
 	}
 	
-	public boolean isEnough(Resource r, float size){
+	public boolean isEnough(Resource r, int size){
 		return resources[r.ordinal()] >= size;
 	}
 
-	public boolean tryRemove(Resource r, float size){
+	public boolean tryRemove(Resource r, int size){
 		if (isEnough(r, size)){
 			resources[r.ordinal()] -= size;
 			return true;
@@ -35,7 +35,7 @@ public class ResourceStorage {
 			return false;
 	}
 	
-	public boolean tryTransfer(Resource r, float size, ResourceStorage rs){
+	public boolean tryTransfer(Resource r, int size, ResourceStorage rs){
 		if (tryRemove(r, size)){
 			rs.add(r, size);
 			return true;
@@ -55,7 +55,7 @@ public class ResourceStorage {
 		}
 	}
 	
-	public void set(Resource r, float size){
+	public void set(Resource r, int size){
 		resources[r.ordinal()] = size;
 	}
 }
